refactor: move row actions out of component
